+++ DomainSeeker_plugin/src/fetch_pdb_pae.py
@@ -1,23 +1,31 @@
-# 全局异常处理
-import domainseeker_errorLog
-#----------------------------------------------------------------------------------------------------
 import os,sys
 import wget
 import numpy as np
 from tqdm import tqdm
 
-UniprotID_list_path=sys.argv[2]
-output_dir=sys.argv[3]
+pae_only = '--pae-only' in sys.argv
+if pae_only:
+    sys.argv.remove('--pae-only')
 
-if len(sys.argv)>4:
-    output_pdb_dir=sys.argv[4]
-    output_pae_dir=sys.argv[5]
+UniprotID_list_path=sys.argv[1]
+output_dir=sys.argv[2]
+
+if pae_only:
+    if len(sys.argv)>3:
+        output_pae_dir=sys.argv[3]
+    else:
+        output_pae_dir=os.path.join(output_dir,'pae_files')
 else:
-    output_pdb_dir=os.path.join(output_dir,'pdb_files')
-    output_pae_dir=os.path.join(output_dir,'pae_files')
+    if len(sys.argv)>3:
+        output_pdb_dir=sys.argv[3]
+        output_pae_dir=sys.argv[4]
+    else:
+        output_pdb_dir=os.path.join(output_dir,'pdb_files')
+        output_pae_dir=os.path.join(output_dir,'pae_files')
 
 
-os.makedirs(output_pdb_dir,exist_ok=True)
+if not pae_only:
+    os.makedirs(output_pdb_dir,exist_ok=True)
 missing_pdbs=[]
 
 
@@ -27,15 +35,16 @@ missing_paes=[]
 
 UniprotID_list=np.loadtxt(UniprotID_list_path,dtype=str,ndmin=1)
 n=len(UniprotID_list)
-for i in tqdm(range(n),desc='Downloading',file=sys.stdout):
+for i in tqdm(range(n),desc='[STATUS]Downloading',file=sys.stdout):
     ID=UniprotID_list[i]
-    pdb_path=os.path.join(output_pdb_dir,ID+'.pdb')
-    if not os.path.exists(pdb_path):
-        pdb_url=f"https://alphafold.ebi.ac.uk/files/AF-{ID}-F1-model_v6.pdb"
-        try:
-            wget.download(pdb_url,pdb_path,bar='')
-        except Exception as e:
-            missing_pdbs.append(ID)
+    if not pae_only:
+        pdb_path=os.path.join(output_pdb_dir,ID+'.pdb')
+        if not os.path.exists(pdb_path):
+            pdb_url=f"https://alphafold.ebi.ac.uk/files/AF-{ID}-F1-model_v6.pdb"
+            try:
+                wget.download(pdb_url,pdb_path,bar='')
+            except Exception as e:
+                missing_pdbs.append(ID)
 
     pae_path=os.path.join(output_pae_dir,ID+'.json')
     if not os.path.exists(pae_path):
@@ -46,12 +55,13 @@ for i in tqdm(range(n),desc='Downloading',file=sys.stdout):
             missing_paes.append(ID)
 
 
-missing_pdb_log_path=os.path.join(output_dir,'missing_pdb.log')
-missing_pdb_log=open(missing_pdb_log_path,'w')
-if len(missing_pdbs)>0:
-    for ID in missing_pdbs:
-        missing_pdb_log.write(ID+'\n')
-missing_pdb_log.close()
+if not pae_only:
+    missing_pdb_log_path=os.path.join(output_dir,'missing_pdb.log')
+    missing_pdb_log=open(missing_pdb_log_path,'w')
+    if len(missing_pdbs)>0:
+        for ID in missing_pdbs:
+            missing_pdb_log.write(ID+'\n')
+    missing_pdb_log.close()
 
 missing_pae_log_path=os.path.join(output_dir,'missing_pae.log')
 missing_pae_log=open(missing_pae_log_path,'w')
@@ -61,8 +71,8 @@ if len(missing_paes)>0:
 missing_pae_log.close()
 
 
-
-missing_pdb_log.close()
-missing_pae_log.close()
-
-print("Done fetching PDB and PAE files.")
+if pae_only:
+    print("Done fetching PAE files.", flush=True)
+else:
+    print("Done fetching PDB and PAE files.", flush=True)
+print("=" * 40, flush=True)